+++ mysite/trading/templates/stock/includes/stock_change_summary.html
@@ -1,0 +1,6 @@
+
+<div class="row">
+            <div class="col-12">
+                {{ summary|safe }}
+            </div>
+        </div>
