+++ mysite/trading/templates/stock/includes/stock_charts.html
@@ -1,0 +1,50 @@
+<div class="row">
+    <div class="col-12">
+        <div class="card">
+            <div class="card-header">
+
+                <ul class="nav nav-tabs card-header-tabs" id="chart-tabs" role="tablist">
+
+                    <li class="nav-item">
+                        <a class="nav-link active" href="#today-chart" role="tab" aria-controls="today-chart" aria-selected="true">Intraday</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="nav-link" href="#historic-chart" role="tab" aria-controls="historic-chart" aria-selected="false">Daily</a>
+                    </li>
+
+                </ul>
+            </div>
+
+            <div class="card-body">
+                <div class="tab-content mt-3">
+                    <div class="tab-pane active" id="today-chart" role="tabpanel">
+                        <h3 class="card-title">Intraday Graph</h3>
+                        <div class="card-text">
+
+                            {% if intraday_graph %}
+                            <div>
+                                {{ intraday_graph|safe }}
+                            </div>
+                            {% endif %}
+
+                        </div>
+                    </div>
+
+                    <div class="tab-pane" id="historic-chart" role="tabpanel">
+                        <h3 class="card-title">Historic Graph</h3>
+                        <div class="card-text">
+
+                            {% if historic_graph %}
+                            <div>
+                                {{ historic_graph|safe }}
+                            </div>
+                            {% endif %}
+
+                        </div>
+                    </div>
+
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
