+++ mysite/trading/templates/stock/includes/stock_list.html
@@ -1,0 +1,34 @@
+<table class="table">
+    <tr>
+        <th scope="col">Ticker</th>
+        <th scope="col">Name</th>
+        <th scope="col">Exchange</th>
+        <th scope="col">Price</th>
+        <th scope="col" class="text-center">Bookmark</i></th>
+    </tr>
+    <tbody>
+    {% for stock in stocks %}
+    <tr>
+        <th scope="row"><a href="{% url 'equity' stock.id %}">{{ stock.ticker }}</a></th>
+        <td>{{ stock.name }}</td>
+        <td>{{ stock.exchange }}</td>
+        <td>
+            {% if stock.current_price %}
+                <p>{{ stock.current_price.close }} | <small>{{ stock.current_price.timestamp }}</small></p>
+            {% endif %}
+        </td>
+
+        <td class="text-center">
+
+            {% if stock in is_favourite %}
+            <a href="{% url 'favourite_stock' stock.id %}"><i class="fas fa-bookmark fa-2x"></i></a>
+            {% else %}
+            <a href="{% url 'favourite_stock' stock.id %}"><i class="far fa-bookmark fa-2x"></i></a>
+            {% endif %}
+
+        </td>
+
+    </tr>
+    {% endfor %}
+    </tbody>
+</table>
